fix(cli): Take in to accound blend size when creating edge COGs (#765)
If a COG is on the border of the source imagery then the cutline blending can interfere if it is
truncated to the same boundary line. We need to add padding that is at least as wide as the blend
value to avoid alpha artefacts of these edges. The value used will include 100 extra pixels to allow
adjancent imagery to be included in edge calculation data.