feat(cogify): output single URL in cogify config command BM-822 (#2899)
#### Description
*What does this change aim to achieve?*

Tweaks the `cogify config` command to write out just the URL to a file
for consumption by Argo, rather than writing out a JSON array which
contains one item.

Slight refactor to the `initImageryFromTiffUrl` function (renamed from
`imageryFromTiffUrl`), to improve ease of calling it directly. This
involves changing the signature to allow passing a
`ConfigProviderMemory` instance into the function, which was previously
handled in `initConfigFromUrls` (which calls `initImageryFromTiffUrl`).

#### Intention
*Why is this change being made? What implications or other
considerations are there?*

This change will allow the Argo workflow to display just the URL without
the surrounding `["..."]` JSON array characters. This will make it
easier for users to navigate to the URL to view the imagery.

By exporting the config created by `initImageryFromTiffUrl`, to access
the imagery you will need to load the specific layer, as we no longer
create an `aerial` tileset, which is what the frontend loads by default
unless a layer is specified via `?i=...`. We may want to consider
refactoring the frontend so if it loads a config with only one layer, it
defaults to show that, without requiring specifying the layer name via
`?i=...`.

#### Checklist
*If not applicable, provide explanation of why.*
- [ ] Tests updated - N/A, CLI command does not have tests.
- [ ] Docs updated - N/A, no existing docs.
- [x] Issue linked in title