fix(cli): Run npm install in the basemaps-config to install the @linz/style (#2900)
### Why?
When we make changes to config file in the basemaps-config repo, we need
to format it before commit. Current way of using prettier is broken
because we need to install @linzjs/style into the cli docker container.
A better solution is just run prettier in the config directory before
commit the changes.

#### Issue Link:
[BM-810](https://toitutewhenua.atlassian.net/browse/BM-810)

### Description of Changes:
- Remove all the prettier stuff from cli
- Run npx prettier in the chirld process in config repo.

#### Author Checklist:
- [ ] Tests updated
- [ ] Docs updated
- [ ] Issue linked in PR title

[BM-810]:
https://toitutewhenua.atlassian.net/browse/BM-810?atlOrigin=eyJpIjoiNWRkNTljNzYxNjVmNDY3MDlhMDU5Y2ZhYzA5YTRkZjUiLCJwIjoiZ2l0aHViLWNvbS1KU1cifQ