feat(cogify): ensure cogify path-like args have trailing slashes. BM-858 (#2903)
#### Description
*What does this change aim to achieve?*

Updates the cogify CLI commands to ensure that arguments representing
folder paths have trailing slashes. Adds a `UrlFolder` parser for cmd-ts
which parses the input string as a URL, and appends a trailing slash if
it's missing. Changing the type from `string` to `URL` (via the `Url` or
new `UrlFolder` parsers) involved changing places that use these values
to construct other paths. Where these are passed to `fsa` functions,
these are converted back to strings using `urlToString`.

#### Intention
*Why is this change being made? What implications or other
considerations are there?*

Parts of the cogify commands were using strings for paths, while other
parts were using URLs. This led to a situation where a string path was
created by `fsa.joinAll()` which did not have a trailing slash, which
was then passed to the URL constructor, resulting in a URL one directory
higher than intended being created, which prevented finding the STAC
collection.json file.

The calls to `urlToString` will be unnecessary once we upgrade to chunkd
v11 which swaps to receiving URLs directly, however that is a much
larger piece of work so for the time being we'll continue to gradually
migrate to using URLs and use `urlToString` as needed when passing
values to `fsa` functions.

#### Checklist
*If not applicable, provide explanation of why.*
- [ ] Tests updated, N/A: no existing tests to update.
- [x] Docs updated
- [x] Issue linked in Title