feat(landing): Order the categories in the layer dropdown BM-880 (#2916)
#### Description

Orders the categories in the layer selector dropdown on the landing
page, so that the new 'Scanned Aerial Imagery' category appears after
Urban and Rural layers.

#### Intention

Currently the order of categories in the dropdown is based on the order
in which they appear in the WMTS file which the landing page loads the
layers from. This could change over time as new layers with different
alphabetical order are added.

#### Checklist

- [ ] Tests updated
- [ ] Docs updated
- [ ] Issue linked in Title