feat(lambda-tiler): create preview images for og:image BM-264 (#2921)
#### Description

Adds a new endpoint `/v1/preview` which generates `og:image` previews
for a specific location.

```
/v1/preview/:tileSet/:tileMatrixSet/:z/:lon/:lat.:tileType
```

example:

```
/v1/preview/aerial/WebMercatorQuad/177.3998405/-39.0852555.webp
```


#### Intention

When people share links to basemaps they are currently presented with
the same image of Lyttelton, which while really nice looking is not
representative of what they are actually get a link to.

This is the first step in providing an actual image of the area people
are linked to, I will add a screenshot tester of this too.



#### Checklist
*If not applicable, provide explanation of why.*
- [ ] Tests updated
- [ ] Docs updated
- [x] Issue linked in Title