fix(cogify): validate that we have access to all the files before starting (#2912)
#### Description

As we download multiple files at the same time, the first collection of
files will often fail with `403 Permission Denied` we will then attempt
to find a new role to use with these files, but as we are downloading
10+ files at the same time, all of these will attempt to find a new role
and assume it.

This can lead to a massive amount of role assumptions while trying to
access files

#### Intention

This adds a check before we start processing imagery, to validate that
we have read access to all the remote hosts by attempting to head one
file from each remote host.


#### Checklist
*If not applicable, provide explanation of why.*
- [ ] Tests updated
- [ ] Docs updated
- [ ] Issue linked in Title