feat(config): detect empty images and ignore them (#2915)
#### Description

Sometimes when tiffs are create they are created outside the bounds of
where the source tiff actually has data.

`gdal_translate` will then create a completly empty tiff, as the tiffs
are created with `sparse_ok=true` the internal tiff tag `TileOffsets`
which contains the file offsets for all the internal tiles will contain
offsets of `0` with byte length of `0`.

If all the offsets are `0` then there is no data inside the tiff.

#### Intention

We are seeing a large number of empty tiffs being imported into our
configuration and until we can clean out the source locations it is
easier to check the tiff to see if it contains data, when we are
creating the configuration.

I have tested this with the current
[linz/basemaps-config](https://github.com/linz/basemaps-config) configs
and it found 4,548 empty tiff files ranging in size from 1KB to 250KB.

#### Checklist
*If not applicable, provide explanation of why.*
- [x] Tests updated
- [ ] Docs updated
- [ ] Issue linked in Title

---------

Co-authored-by: Andrew Jacombs <ajacombs@linz.govt.nz>