feat: add og:image preview to all basemaps links BM-264 (#2925)
#### Description

Opengraph allows third party websites/applications to embed useful
information about your page into their page.

Currently basemaps serves a default image to all of these applications
https://basemaps.linz.govt.nz/basemaps-card.jpeg which is a nice picture
of Lyttleon rather than what is actually viewed.

Adding a picture that is a preview of what the link displays is very
useful with tools like slack, where often I take a screen shot of the
page and then paste a link with the screenshot.

Example Links

#### Slack

![image](https://github.com/linz/basemaps/assets/1082761/49752243-e407-45f7-bc43-4b432d071c1e)

#### LinkedIn 


![image](https://github.com/linz/basemaps/assets/1082761/bd682b4b-c1f2-4a11-9190-8f1d456c0541)

#### Intention

Because the location of the page is stored only in a hash `#@...` this
is never sent to the server so the server cannot update the
`index.html`'s `og:image` location.

This change modifies the default URL from `#@...` to `/@` which converts
it from a hash to a path. The path is then sent to the server which
processes it updates the latest index.html and serves it back to the
client.

The current lambda function does not have direct access to the
`index.html` so it has to fetch it from `s3://` if this fetch fails, the
user is just redirected back to the landing page. This could be improved
in the future by letting the lambda contain a `index.html` but it breaks
the deployment separation of the landing/tile server.


#### Checklist
*If not applicable, provide explanation of why.*
- [x] Tests updated
- [ ] Docs updated
- [x] Issue linked in Title