feat(lambda-tiler): use the same checker background that the landing does for previews BM-264 (#2929)
#### Description
Adds a checkboard background to all preview images that have transparent
backgrounds

![z8_177 3998405_-39
0852555](https://github.com/linz/basemaps/assets/1082761/78efe50f-47e3-463b-85a5-841a7753f61e)


#### Intention

Currently the preview generator uses a slightly pink background for all
transparency this is not consistent with the landing page.

To be more consistent create a checkerboard background of the same color
to the landing page.

The tiles are slightly smaller 30px vs 50px so they align with the
1200px x 630px dimensions of the preview image




#### Checklist
*If not applicable, provide explanation of why.*
- [x] Tests updated
- [ ] Docs updated
- [x] Issue linked in Title