fix(config): ignore the argo folder "flat/" in guessing imagery names (#2939)
#### Description

When [topo-workflows](https://github.com/linz/topo-workflows) processes
imagery, it puts all the resulting imagery tiffs into `flat/`, when we
attempt to create a config from this folder it names the imagery "flat"
this name is not very helpful for determining what the imagery is of.

#### Modification

Ignores `flat/` from folder names when guessing imagery names.