Fix(cli): Import config to get target zoom from covering.json (#2947)
#### Description
The new cogify workflow deprecate the old job.json and use covering.json
to prepare jobs. We need fix the import to be able to support both
job.json and covering.json.


#### Intention
We try to get gsd from job.json first, if not exist we loop for
covering.json. Also, I have moved the prepare url inside the if
condition to stop these s3 read and only try to get job.json if there is
a change in config.



#### Checklist
- [ ] Tests updated
- [ ] Docs updated
- [ ] Issue linked in Title