fix(tiler-sharp): do not multiply imagery with the background color BM-885 (#2953)
#### Description

Some imagery is being combined with the background image causing white
lines around the imagery when a background color is used.



![dce9ed_22_4173290_2576669](https://github.com/linz/basemaps/assets/1082761/3bbbbbae-6e78-4c10-80fa-33888b69f9ec)

When using a transparent background the problem goes away


![transparent_22_4173290_2576669](https://github.com/linz/basemaps/assets/1082761/4f39935e-d5da-44f4-a79d-e709edbd43e4)

This is caused by the imagery being mixed with the background as part of
the composition process.


#### Intention

`ret.info.*` contains extra metadata which can change how imagery is
overlayed into the output file, by only allowing the metadata we want
`height`, `width` and `channels` this restricts the output image
creation process to how we want.


#### Checklist
*If not applicable, provide explanation of why.*
- [ ] Tests updated
- [ ] Docs updated
- [x] Issue linked in Title