fix(cogify): ignore existing overviews so they are not recompressed (#2954)
#### Description

To create the highest quality overviews, tell GDAL to create new
overviews directly from the source imagery


#### Intention

Because we are mostly importing RGBA imagery from webp COGs now instead
of straight geotiffs, overviews now exist and GDAL is starting to
automatically use them.

The overviews are generally lossly already so using them as a base to
create more overviews means we are compounding the lossy compression.


#### Checklist
*If not applicable, provide explanation of why.*
- [ ] Tests updated
- [ ] Docs updated
- [ ] Issue linked in Title