feat: Add proof of concept docs site BM-917 (#2990)
#### Motivation

This adds an initial proof of concept of a documentation site. The
documentation is generated from a combination of static Markdown
documents; currently lightly modified from the existing files, plus a
few placeholders.

API documentation is generated from the typescript source code using
Typedoc. This is rendered to Markdown files using
typedoc-plugin-markdown, which is combined with the static Markdown
files to render the documentation site using MkDocs.

#### Modification

This is a rough proof of concept implementation, to be fleshed out in
subsequent PRs which will:
- Add more documentation.
- Improve the readability of the generated API docs (fix typographic
hierarchy and character escaping issues).
- Update the theme to match LINZ branding.

#### Checklist

- [ ] Tests updated
- [x] Docs updated
- [x] Issue linked in Title