fix(cogify): prevent empty tiffs from being stored (#3018)
#### Motivation

As cogify does not have the exact bounding polygon of the data inside
the tiff, there are occasions where a tiff is created outside the extent
of the data. this tiff is empty and gdal will create a empty sparse
tiff.

These tiffs are useless and should be ignored/deleted.


#### Modification

Check the output tiff to see if its empty before uploading it.

#### Checklist

_If not applicable, provide explanation of why._

- [ ] Tests updated
- [ ] Docs updated
- [ ] Issue linked in Title