feat(config): disable the empty tiff checker (#3023)
#### Motivation

The empty tiff checker was designed to remove empty tiffs from the
configuration as we had 9,000 tiffs that were empty. It greatly slows
down the build pipeline.

We have since deleted the tiffs and have added logic to prevent them
being uploaded in future.


#### Modification

remove the `isEmpty` check when creating config from tiffs.


#### Checklist

_If not applicable, provide explanation of why._

- [ ] Tests updated
- [ ] Docs updated
- [ ] Issue linked in Title