feat: allow configuration of output tile types base of tileset configuration BM-932 (#3103)
#### Motivation

when rendering more complex raster data types there needs to be a
configuration on how to convert the source raster images into a RGBA
output to be consumed. This is needed so that DEM and DSM can be
configured into TerrainRGB and other formats such as color ramps.

#### Modification

Allows configuration of how output tiles are created on a per tileset
basis

#### Checklist

_If not applicable, provide explanation of why._

- [ ] Tests updated
- [ ] Docs updated
- [ ] Issue linked in Title

---------

Co-authored-by: Wentao Kuang <wkuang@linz.govt.nz>