docs: Document the gdal command to re-project gebco into 2193. (#3098)
#### Motivation

We need process gebco data into basemaps once a year once gebco get new
updates. But there is a bug in cogify (BM-935) when deal with
NZTM2000Quad on Wsg84 source within the whole world scale, as the out
put extent is not clipped as expected multi-polygons.

This bug is only happens when the source file is too large that crossing
the whole world, and existing the logic is very complex, so if trying to
fix it might have a high risk to broke some existing process. As a
result, we decide to using gdal to reproject he gebco into 2193 before
using cogify importing the data.

This PR contains the documents of dgal commands that using to re-project
gebco data.

#### Modification

Add docs for the gdal commands.

#### Checklist

_If not applicable, provide explanation of why._

- [ ] Tests updated
- [ ] Docs updated
- [ ] Issue linked in Title

---------

Co-authored-by: Blayne Chard <bchard@linz.govt.nz>