fix(cli): Fix the missing format for vector config with Basemaps v7. (#3149)
#### Motivation
We have done some refactoring for the config structure to support
elevation data in future, but this moved the `format` into
`output.format`. This breaking change will broken the vector config in
the current basemaps version. We should make sure the transaction of
format stay consistent.

#### Modification

Add a optional `format` for the `ConfigTileSetVector`

#### Checklist

_If not applicable, provide explanation of why._

- [ ] Tests updated
- [ ] Docs updated
- [ ] Issue linked in Title