fix(cogify): improve DEM quality when reprojecting and scaling BM-987 (#3189)
#### Motivation

Baseamps generally uses three GDAL commands to convert imagery/DEM/DSMs
into COGS

build a VRT
`gdalbuildvrt tiffs.vrt *.tiff`

Apply cutline and warp
`gdalwarp -of vrt warped.vrt tiffs.vrt -wo ... `

Convert the warped to a COG

`gdal_translate -of COG warped.vrt`

When reprojecting we often change the resolution of a TIFF to better
align with the desired output tile matrix this is done with a `-tr`
inside the `gdal_translate` step.

We have found that with just a single `-tr` in the `gdal_translate` it
leads to suboptimal COG creation with artifacts.


![old](https://github.com/linz/basemaps/assets/1082761/ab66aeff-7047-4968-b9ad-dec2041ae022)


adding another `-tr` to the warp resolves the issue.


![new](https://github.com/linz/basemaps/assets/1082761/2627537e-7db8-4992-8795-702229fa3da1)


#### Modification

- Adds `-tr` to the gdalwarp to fix some rendering issues
- Adds a `cogify validate` to validate tiffs are what basemaps is
expecting.
- 
#### Checklist

_If not applicable, provide explanation of why._

- [ ] Tests updated
- [ ] Docs updated
- [ ] Issue linked in Title