feat(landing): exaggerate the terrain when viewing in NZTM BM-983 (#3199)
#### Motivation

When viewing terrain in NZTM the terrain is too flat, this is caused by
the NZTM projection hacks to allow Maplibre to render NZTM, it offsets
the zoom level by 2 which means everything is approx 4x smaller.

#### Modification

When viewing in NZTM set the terrain to be approx 4x higher than what we
use in 3857.

#### Checklist

_If not applicable, provide explanation of why._

- [ ] Tests updated
- [ ] Docs updated
- [ ] Issue linked in Title