fix(lambda-tiler): ensure wmts limits extent to the bounding box of the tile matrix extent BM-1012 (#3235)
#### Motivation

As we recently changed the basemaps configuration to use the raw tiff
values for extent rather than the name of the tile, the extents of each
layer has slightly changed.

This works for most layers except ones which cover the entire world, for
example `0-0-0.tiff` this can lead to slight floating point errors in
the extents between the TileMatrix and the tiff.

#### Modification

If the imagery extent is outside the bounds of the tile matrix fall back
to the extent of the tile matrix.

#### Checklist

_If not applicable, provide explanation of why._

- [ ] Tests updated
- [ ] Docs updated
- [ ] Issue linked in Title

---------

Co-authored-by: Wentao Kuang <wkuang@linz.govt.nz>