fix(cogify): correct loading path for "cogify" bin (#3294)
### Motivation

the latest container is currently broken when looking for `cogify` as it
is pointing to a file that doesnt exist `./dist/index.cjs`.

The dist/index.cjs is only created after a `npm run bundle` but bundle
is not run until after cogify is published to npm.

<!-- TODO: Say why you made your changes. -->

### Modifications

force dist/index.cjs to be created before a `npm pack` can be run using
the prepack npm lifecylce event.

<!-- TODO: Say what changes you made. -->

<!-- TODO: Attach screenshots if you changed the UI. -->

### Verification

<!-- TODO: Say how you tested your changes. -->

manually rebuilt with `npm pack` to ensure that a dist/index.cjs exists