feat(cogify): error early if no source collection.json is found BM-1047 (#3296)
### Motivation

When importing imagery into basemaps we know if there is a source stac
collection at the first step, but we only warn that we cannot find it.

In most of our import processes not having a stac collection should be a
errro

<!-- TODO: Say why you made your changes. -->

### Modifications

Add `--require-stac-collection` to throw a error if a stac collection
cannot be found when createing a cover

<!-- TODO: Say what changes you made. -->

<!-- TODO: Attach screenshots if you changed the UI. -->

### Verification

<!-- TODO: Say how you tested your changes. -->

Added new integration test to covering creation with and without the new
flag.