feat(landing): Add LINZ-Terrain-Prod for debug 3d map with production elevation data. BM-1058 (#3323)
### Motivation

We want debug page to view terrain for individual config files that
doesn't have elevation tileset inside. Using the same logic of
linz-aerial slider and get the elevation data from the production or dev
config based on the host.

### Modifications

Add a new LINZ-Terrain-Prod source layer in the terrain and hillshade
dropdown that points to the basemaps production data for debugging.

### Verification

![image](https://github.com/user-attachments/assets/a8888c20-8e52-4c00-b5c2-f938b3c7a3d3)