fix(landing): Fix the map switcher that setting bearing and pitch to 0. BM-1063 (#3326)
### Motivation
Landing page map switcher on the bottom left doesn't set the map
location when bearing or pitch is 0.

### Modifications
The map switcher sets the bearing and pitch by the update location from
config. But it is skip setting when pitch or bearing is null. We should
also set to 0 if null.

### Verification

---------

Co-authored-by: Blayne Chard <bchard@linz.govt.nz>