feat(linzjs-geojson): add iterate and truncate utilities for geojson (#3340)
### Motivation

We have had a need to reproject and truncate geojson objects and instead
of copying the logic into multiple locations add it to the geojson
helper we have here.

### Modifications

Add truncate to truncate lat lon pairs to by default 8 decimal places
Add iterate to iterate over all the points in a geojson feature

### Verification

unit tests and is currently in use in other code bases.