feat(landing): Add new features pop up to alert users for recent releases. BM-1064 (#3356)
### Motivation
We want to alert Basemaps users about new features for recent release.
To use the @linzjs/lui ->
[LuiUpdatesSplashModal](https://github.com/linz/Lui/blob/master/src/components/LuiUpdateSplashModal/LuiUpdatesSplashModal.tsx)
which is same with land online search.

### Modifications
Initial investigations with @linzjs/lui that it's very hard to import
and use react components from it because different tech stacks.
- lui build and bundled with nodejs which contains dependencies like
`crypto` or `stream` that is not support by browser. We might need to do
polyfills to map them into browser friendly dependencies.
- lui only bundle everything into one index.js, we don't have individual
module file to do direct import, so we have to deal with dependency
problem above.

As a result, we decide to re-implement the `LuiUpdatesSplashModal` in
basemaps just using the assets and css from @linzjs/lui
- `FeatureUpdates` is a completely re-implementation of
`LuiUpdatesSplashModal`
- `NewFeature` is the component that we need to update often when
describe new features.
- Updated @linzjs/lui to latest

### Verification
- Tested new pop up only show once after dismissed.
- Tested new pop up stop showing after closing date.
- Tested new pop up shows in incognito mode/clearing cache after
dismissed.

![image](https://github.com/user-attachments/assets/2440883b-9b00-4bf1-bfdc-a00de6f21fc3)

![image](https://github.com/user-attachments/assets/7061711b-81a5-407c-b442-ea02ef8bf161)