fix(lambda-tiler): correctly log fetch requests (#3359)
### Motivation

Requests tracing has been broken for a while (not entire sure when it
broke)

It is very very helpful to have statistics on the number of s3 requests
required to make a tile

### Modifications

- `requestCount` is reserved for the number of lambda invocations so
move to `fetchCount` for number of fetches
- include the fetchIds to trace hot fetches
- increase the percent of requests with tracing enabled
- Use async local storage to gain access to the logger for the request
that triggers the log message, this fixes a bug where the request may be
in `trace` but the fetch is using the base logger which is set to `info`
so no logs are generated.


### Verification

Ran locally and verified logs are now appearing.

<!-- TODO: Say how you tested your changes. -->

---------

Co-authored-by: Wentao Kuang <wkuang@linz.govt.nz>