fix(lambda-tiler): NZTM terrain been multiplied twice. BM-1122 (#3373)
### Motivation

Basemap terrain exaggeration for NZTM been multiplied twice.

![image](https://github.com/user-attachments/assets/4e7b4a77-431c-4f0e-be5d-206787ae2452)

![image](https://github.com/user-attachments/assets/e8a64789-7e70-4bb9-97f8-c8cac4e06e36)

### Modifications
We have manually multiplied the NZTM terrain exaggeration when
converting the style json earlier, however, we got default setting to
fix them as 4.4 in `DefaultExaggeration`, so we don't need to covert the
exaggeration twice to make it too large.

### Verification
Local test and unit test.

![image](https://github.com/user-attachments/assets/42f35f6f-6410-4ccd-87f6-c007f9faae7f)