fix(landing): Fix the zoom to extend by using maplibre to adjust AntiMeridian. BM-1155 (#3384)
### Motivation

The zoom to extend feature is broken after upgrade the maplibre to
latest release. It always zoom back to the whole world when switching
layer, seems the caused by the anti Meridian. And this only happens to
the Web Mercator

### Modifications
Maplibre got a fix of
[adjustAntiMeridian](https://github.com/maplibre/maplibre-gl-js/blob/87486a5ef2085e600e8fa4e31252629dd8488dcd/src/geo/lng_lat_bounds.ts#L342)
between v4.5 to v4.71.
I believe that causes some conflicts with the pre-adjust in our system,
and this is fixed after we do a `maplibre.LngLatBounds.convert` for Web
Mercator as well.

### Verification
Local tested, all zoom to layer working fine for Web Mercator and NZTM
now.