feat: improved cloudwatch analytics rollup BM-1092 (#3381)
#### Motivation

We currently get 10s of millions of requests per day, which are all
logged by cloudwatch into a s3 bucket, these logs are very large when
looking over the entire history of basemaps, so we rollup the logs into
hourly summaries of the requests.

The current implementation rolls up the logs into coarse buckets these
buckets give a good overview of usage but do not allow us to dig into
more detailed usage trends such as which datasets are used in what
formats.

### Modification

Introduces a new analytic lambda to run side by side with the existing
lambda until we are comfortable to turn off the old lambda.

The logs are now rolled aggregated on more fields such as 
- tiles
  - zoom level
  - matrix set (and requested matrix set, /3857/ vs /WebMercatorQuad/ )
  - extension, png vs pbf
- tile set
- api key and type

---------

Co-authored-by: Wentao Kuang <wkuang@linz.govt.nz>