feat(shared): support multiple credential locations (#3407)
### Motivation

We have a number of configuration files to allow reading and writing of
S3 buckets across accounts, sometimes it's useful to join a collection
of access lists together to provide a more comprehensive list of
credentials for the calling service.

### Modifications

Allow credentials to be loaded from multiple sources using either a
comma separated list or a JSON list, this behavior is the same as the
workflows inside of argo

https://github.com/linz/argo-tasks/blob/master/src/fs.register.ts#L132

### Verification

Logic is taken from argo tasks, and will be tested in argo.