feat: CITM2000 debug tile matrix (#3397)
### Motivation

We are starting to want to view imagery in other projections that do not
have a standard Web Quad tile matrixes to work with. It would be helpful
to have a way to generate a debug tile matrix that aligns with the
WebMercatorQuad standard scales, while also being a quad starting at 1x1
-> 2x2 -> 4x4 grids that increase with powers of two.

### Modifications

Added a CITM Debug tile matrix based off the WebMercatorQuad scales.

Added a script to generate tile matrices from a supplied EPSG
projection,

The script will download the projection extent from a projjson, then
calculate an appropriate WebMercatorQuad scale that will cover the
entire projection extent,

This script was only tested with CITM and NZTM it will likely fail if
used with other projections.

### Verification

Tested basemaps server with QGIS and some CI topo50 mapsheets.


![image](https://github.com/user-attachments/assets/6e91b8cf-9525-4e24-8c09-dcd0432107b3)