fix(lambda-analytic-cloudfront): process logs upto one hour ago (#3409)
### Motivation

Log processing is done in batches and is restricted to at most one hour
behind, there is a bug in the existing logic where logs processing can
be delayed by upto a day.

This is caused by the looping logic where it subtracts one day then
loops over the 24 hours of that day, so if the time is close to the day
start (eg 1am utc) the logs are processed upto midnight of the day
before, if the current time is close to the end of the day (eg 11pm utc)
the logs will be processed up to midnight the day before.

### Modifications

Move the processing logic to start looping from one hour ago, 
Exclude any timestamps that are in the future.

### Verification

Unit tests